arm: at91: gardena-smart-gateway-at91sam: Fix udelay() in SPL
This prevents SPL from panicking with:

    Could not initialize timer (err -19)

It appears that the `udelay()` call in `ddr2_init()` has not been
working as expected for a long time. Probably since the introduction of
the DM-based PIT timer. Since commit 4b2be78ab6 ("time: Fix get_ticks
being non-monotonic") this is an actual problem for us, as our board
would no longer boot.

Signed-off-by: Ezra Buehler <ezra.buehler@husqvarnagroup.com>